+++ React ToDo/src/hooks/useTodos.js
@@ -1,0 +1,29 @@
+import { useRef, useState } from 'react'
+
+export function useTodos() {
+  const lastId = useRef(4)
+
+  const [todos, setTodos] = useState([
+    { id: 3, text: '공부하기', checked: true },
+    { id: 2, text: '코딩하기', checked: false },
+    { id: 1, text: '운동하기', checked: true },
+  ])
+
+  const addTodo = (text) => {
+    const todo = { id: lastId.current, text, checked: false }
+    setTodos([todo, ...todos])
+    lastId.current++
+  }
+
+  const removeTodo = (seletedId) => {
+    const filterTodos = todos.filter((todo) => todo.id != seletedId)
+    setTodos(filterTodos)
+  }
+
+  const toggleTodo = (seletedId) => {
+    const updateTodos = todos.map((todo) => (todo.id == seletedId ? { ...todo, checked: !todo.checked } : todo))
+    setTodos(updateTodos)
+  }
+
+  return { todos, addTodo, removeTodo, toggleTodo }
+}
